pyproject.toml
[project]
name = "kloe-bdt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "awkward>=2.9.0",
    "fastapi>=0.136.0",
    "grip>=4.6.2",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "mysql-connector-python>=9.6.0",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "psutil>=7.2.2",
    "pyarrow>=23.0.1",
    "scikit-learn>=1.8.0",
    "scikit-optimize>=0.10.2",
    "seaborn>=0.13.2",
    "uproot>=5.7.1",
    "xgboost>=3.2.0",
]
